Fix PF6 issues for Prviders dropdown menus and modals alignments
Reference:  https://issues.redhat.com/browse/MTV-1982
