+++ locales/en/plugin__forklift-console-plugin.json
@@ -20,7 +20,6 @@
   "{{total}} VM_other": "{{total}} VM",
   "{children}": "{children}",
   "{projectName} project does not contain any providers. To select a source provider, create a provider in this project or select a different project.": "{projectName} project does not contain any providers. To select a source provider, create a provider in this project or select a different project.",
-  "<0>`URL of the Red Hat Virtualization Manager (RHVM) API endpoint.</0><1></1><2>The format of the provided URL of the Red Hat Virtualization Manager (RHVM) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /api, for example: <1>https://rhv-host-example.com/ovirt-engine/api</1>.</2>": "<0>`URL of the Red Hat Virtualization Manager (RHVM) API endpoint.</0><1></1><2>The format of the provided URL of the Red Hat Virtualization Manager (RHVM) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /api, for example: <1>https://rhv-host-example.com/ovirt-engine/api</1>.</2>",
   "<0><0>Choose the root filesystem to be converted.</0><1>Default behavior is to choose the first root device in the case of a multi-boot operating system. Since this is a heuristic, it may sometimes choose the wrong one.</1><2>When using a multi-boot VM, you can also name a specific root device, eg. <2>/dev/sda2</2> would mean to use the second partition on the first hard drive. If the named root device does not exist or was not detected as a root device, the migration will fail. <5>Learn more</5>.</2></0>": "<0><0>Choose the root filesystem to be converted.</0><1>Default behavior is to choose the first root device in the case of a multi-boot operating system. Since this is a heuristic, it may sometimes choose the wrong one.</1><2>When using a multi-boot VM, you can also name a specific root device, eg. <2>/dev/sda2</2> would mean to use the second partition on the first hard drive. If the named root device does not exist or was not detected as a root device, the migration will fail. <5>Learn more</5>.</2></0>",
   "<0><0>Default behavior is to choose the first root device in the case of a multi-boot operating system. Since this is a heuristic, it may sometimes choose the wrong one.</0><1>When using a multi-boot VM, you can also name a specific root device, eg. <2>/dev/sda2</2> would mean to use the second partition on the first hard drive. If the named root device does not exist or was not detected as a root device, the migration will fail.</1><2><0>Learn more</0></2></0>": "<0><0>Default behavior is to choose the first root device in the case of a multi-boot operating system. Since this is a heuristic, it may sometimes choose the wrong one.</0><1>When using a multi-boot VM, you can also name a specific root device, eg. <2>/dev/sda2</2> would mean to use the second partition on the first hard drive. If the named root device does not exist or was not detected as a root device, the migration will fail.</1><2><0>Learn more</0></2></0>",
   "<0><0>MTV behavior is based on the <1>Shared disks</1> setting in the plan.</0><1>If this is set to <1>true</1>, the shared disks will be migrated.</1><2>If this is set to <1>false</1>, the shared disks will not be migrated.</2></0>": "<0><0>MTV behavior is based on the <1>Shared disks</1> setting in the plan.</0><1>If this is set to <1>true</1>, the shared disks will be migrated.</1><2>If this is set to <1>false</1>, the shared disks will not be migrated.</2></0>",
@@ -41,6 +40,7 @@
   "<0>Start the warm migration for plan <1>{name}</1>?</0><1></1><2>VMs included in warm migrations migrate with minimal downtime.</2>": "<0>Start the warm migration for plan <1>{name}</1>?</0><1></1><2>VMs included in warm migrations migrate with minimal downtime.</2>",
   "<0>URL of the Openshift Virtualization API endpoint.</0><1></1><2>The format of the provided URL of the Openshift Virtualization API endpoint should include a scheme, a domain name, and, optionally a port. For example: <2>https://api.openshift-domain.com:6443</2>.</2>": "<0>URL of the Openshift Virtualization API endpoint.</0><1></1><2>The format of the provided URL of the Openshift Virtualization API endpoint should include a scheme, a domain name, and, optionally a port. For example: <2>https://api.openshift-domain.com:6443</2>.</2>",
   "<0>URL of the OpenStack Identity (Keystone) endpoint.</0><1></1><2>The format of the provided URL of the OpenStack Identity (Keystone) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will indicate the server version, for example: <1>https://identity_service.com:5000/v3</1>.</2>": "<0>URL of the OpenStack Identity (Keystone) endpoint.</0><1></1><2>The format of the provided URL of the OpenStack Identity (Keystone) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will indicate the server version, for example: <1>https://identity_service.com:5000/v3</1>.</2>",
+  "<0>URL of the Red Hat Virtualization Manager (RHVM) API endpoint.</0><1></1><2>The format of the provided URL of the Red Hat Virtualization Manager (RHVM) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /api, for example: <1>https://rhv-host-example.com/ovirt-engine/api</1>.</2>": "<0>URL of the Red Hat Virtualization Manager (RHVM) API endpoint.</0><1></1><2>The format of the provided URL of the Red Hat Virtualization Manager (RHVM) API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /api, for example: <1>https://rhv-host-example.com/ovirt-engine/api</1>.</2>",
   "<0>URL of the vCenter API endpoint.</0><1></1><2>The format of the URL of the vCenter API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /sdk, for example: <2>https://vCenter-host-example.com/sdk</2>.</2>": "<0>URL of the vCenter API endpoint.</0><1></1><2>The format of the URL of the vCenter API endpoint should include a scheme, a domain name, path, and optionally a port. Usually the path will end with /sdk, for example: <2>https://vCenter-host-example.com/sdk</2>.</2>",
   "<0>Use the CA certificate of the Manager unless it was replaced by a third-party certificate, in which case enter the Manager Apache CA certificate. You can retrieve the Manager CA certificate at https://[engine_host]/ovirt-engine/services/pki-resource?resource=ca-certificate&format=X509-PEM-CA.</0><1>When left empty the system CA certificate is used.</1><2>The certificate is not verified when <1>Skip certificate validation</1> is set.</2>": "<0>Use the CA certificate of the Manager unless it was replaced by a third-party certificate, in which case enter the Manager Apache CA certificate. You can retrieve the Manager CA certificate at https://[engine_host]/ovirt-engine/services/pki-resource?resource=ca-certificate&format=X509-PEM-CA.</0><1>When left empty the system CA certificate is used.</1><2>The certificate is not verified when <1>Skip certificate validation</1> is set.</2>",
   "<0>Use the CA certificate of the Manager unless it was replaced by a third-party certificate, in which case enter the Manager Apache CA certificate.</0><1>When left empty the system CA certificate is used.</1><2>The certificate is not verified when <1>Skip certificate validation</1> is set.</2>": "<0>Use the CA certificate of the Manager unless it was replaced by a third-party certificate, in which case enter the Manager Apache CA certificate.</0><1>When left empty the system CA certificate is used.</1><2>The certificate is not verified when <1>Skip certificate validation</1> is set.</2>",
