+++ src/modules/Providers/modals/EditModal/EditModal.style.css
@@ -1,12 +1,4 @@
-.forklift-edit-modal-body {
-  margin: 0;
-  padding-bottom: var(--pf-global--spacer--md);
-}
 
-.forklift-edit-modal-form {
-  margin: 0;
-  padding-bottom: var(--pf-global--spacer--md);
-}
 
 .forklift-edit-modal-field-error-validation {
   color: var(--pf-c-form__helper-text--m-error--Color);
