+++ src/providers/components/OpenshiftNetworkCell.tsx
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom-v5-compat';
 import { getResourceFieldValue } from 'src/components/common/FilterGroup/matchers';
 import { TableEmptyCell } from 'src/modules/Providers/utils/components/TableCell/TableEmptyCell';
 import { TableLabelCell } from 'src/modules/Providers/utils/components/TableCell/TableLabelCell';
-import { getResourceUrl } from 'src/modules/Providers/utils/helpers/getResourceUrl';
 import type { CellProps } from 'src/modules/Providers/views/list/components/CellProps';
 
-import { ProviderModelRef } from '@kubev2v/types';
 import { NetworkIcon } from '@patternfly/react-icons';
+
+import { getProviderURL } from '../utils/getProviderURL';
 
 type OpenshiftNetworkCellProps = {
   inventoryValue?: number;
@@ -23,11 +23,7 @@ const OpenshiftNetworkCell: FC<OpenshiftNetworkCellProps> = ({
   const value = fields?.length
     ? getResourceFieldValue({ ...provider, inventory }, fieldId, fields)
     : inventoryValue;
-  const providerURL = getResourceUrl({
-    name: provider?.metadata?.name,
-    namespace: provider?.metadata?.namespace,
-    reference: ProviderModelRef,
-  });
+  const providerURL = getProviderURL(provider);
 
   if (value === undefined) {
     return <TableEmptyCell />;
